+++ bindings/matrix-sdk-crypto-js/src/verification.rs
@@ -2,11 +2,13 @@
 
 #[cfg(feature = "qrcode")]
 use std::fmt;
+use std::time::Duration;
 
 use futures_util::StreamExt;
 #[cfg(feature = "qrcode")]
 use js_sys::Uint8ClampedArray;
 use js_sys::{Array, Function, JsString, Promise};
+use matrix_sdk_common::instant::Instant;
 use matrix_sdk_crypto::VerificationRequestState;
 use ruma::events::key::verification::{
     cancel::CancelCode as RumaCancelCode, VerificationMethod as RumaVerificationMethod,
@@ -906,6 +908,20 @@ impl VerificationRequest {
         self.inner.timed_out()
     }
 
+    /// The number of milliseconds remaining before this verification flow times
+    /// out.
+    ///
+    /// Returns zero if the time has already passed.
+    #[wasm_bindgen(js_name = "timeRemainingMillis")]
+    pub fn time_remaining_millis(&self) -> f64 {
+        (self
+            .inner
+            .timeout_time()
+            .checked_duration_since(Instant::now())
+            .unwrap_or(Duration::from_secs(0)))
+        .as_millis() as f64
+    }
+
     /// Get the supported verification methods of the other side.
     ///
     /// Will be present only if the other side requested the
@@ -1073,6 +1089,8 @@ impl VerificationRequest {
     }
 
     /// Transition from this verification request into a SAS verification flow.
+    ///
+    /// Returns `Promise<[Sas, RoomMessageRequest|ToDeviceRequest] | undefined>`
     #[wasm_bindgen(js_name = "startSas")]
     pub fn start_sas(&self) -> Promise {
         let me = self.inner.clone();