+++ bindings/matrix-sdk-crypto-js/CHANGELOG.md
@@ -5,6 +5,8 @@
 -   In `OlmMachine.getIdentity`, wait a limited time for any in-flight
     device-list updates to complete.
 
+-   Add `VerificationRequest.timeRemainingMillis()`.
+
 # v0.1.0-alpha.11
 
 ## Changes in the Javascript bindings